+++ android/src/main/java/com/pspdfkit/flutter/pspdfkit/FlutterInstantPdfActivity.kt
@@ -39,7 +39,6 @@ class FlutterInstantPdfActivity : InstantPdfActivity() {
     
     override fun onDocumentLoaded(pdfDocument: PdfDocument) {
         super.onDocumentLoaded(pdfDocument)
-        EventDispatcher.getInstance().notifyDocumentLoaded(pdfDocument)
         val result = loadedDocumentResult.getAndSet(null)
         result?.success(true)
     }