+++ android/src/main/java/com/pspdfkit/flutter/pspdfkit/FlutterPdfActivity.java
@@ -60,7 +60,6 @@ public class FlutterPdfActivity extends PdfActivity {
     @Override
     public void onDocumentLoaded(@NonNull PdfDocument pdfDocument) {
         super.onDocumentLoaded(pdfDocument);
-        EventDispatcher.getInstance().notifyDocumentLoaded(pdfDocument);
         Result result = loadedDocumentResult.getAndSet(null);
         if (result != null) {
             result.success(true);